LayoutTests/fast/regions/cssom/get-region-flow-ranges-expected.txt

[CSSRegions] Test for Region::getRegionFlowRanges.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Text node.
Empty regions
PASS range1.length is 0
PASS range2.length is 0
PASS range3.length is 0
Flow article
PASS range1.length is 1
PASS Array [wrapper,0,box1,1] is equal to [wrapper,0,box1,1]
PASS range2.length is 1
PASS Array [box1,1,box1,1] is equal to [box1,1,box1,1]
PASS range3.length is 1
PASS Array [box1,1,wrapper,1] is equal to [box1,1,wrapper,1]
Add text node to region2 (creates anonymous block)
PASS range1.length is 1
PASS Array [wrapper,0,box1,1] is equal to [wrapper,0,box1,1]
PASS range2.length is 1
PASS Array [box1,1,box1,2] is equal to [box1,1,box1,2]
PASS range3.length is 1
PASS Array [box1,2,wrapper,1] is equal to [box1,2,wrapper,1]
Increase font size
PASS range1.length is 1
PASS Array [wrapper,0,box1,1] is equal to [wrapper,0,box1,1]
PASS range2.length is 1
PASS Array [box1,1,#text,4] is equal to [box1,1,#text,4]
PASS range3.length is 1
PASS Array [#text,5,wrapper,1] is equal to [#text,5,wrapper,1]
Make child a content node, so it will move from region1 to region3
PASS range1.length is 2
PASS Array [wrapper,0,box1,0] is equal to [wrapper,0,box1,0]
PASS Array [box1,1,#text,4] is equal to [box1,1,#text,4]
PASS range2.length is 1
PASS Array [#text,5,box1,2] is equal to [#text,5,box1,2]
PASS range3.length is 1
PASS Array [box1,0,box1,1] is equal to [box1,0,box1,1]
PASS successfullyParsed is true

TEST COMPLETE
